requirements.txt
# ==============================================================================
# Project VERA - Dependencies
# Virtual Engineering Review Agent
# ==============================================================================

# --- Core LLM & Agent Framework ---
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langchain-ollama>=1.0.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# --- Embeddings ---
sentence-transformers>=3.0.0

# --- Vector Store ---
chromadb>=0.5.0

# --- Utilities ---
python-dotenv>=1.0.0
streamlit>=1.40.0

# --- Document Generation ---
python-docx>=1.1.0
